requirements.txt

Flask==3.0.3
joblib==1.4.2
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
surprise==0.1
tzdata==2024.2
waitress==3.0.0
Werkzeug==3.0.4
